Collect plugins used in Java compilation and pass them to android_lint.
Android linter needs those plugins. They are not the same as the ones returned in JavaInfo. For example `java_library(... plugins = [x] ...)` would use x in the compilation, but not return it in its own JavaInfo.

The change introduces a slight RAM regression, which cannot be removed until middle layer of Java is Starlarkfied (i.e. java_common.compile creates this array, but we can't reuse it in Starlark at the moment).

PiperOrigin-RevId: 387535200