bzlmod: Store a Package instead of a Rule in BzlmodRepoRuleValue
(https://github.com/bazelbuild/bazel/issues/13316)

This is to work around a restriction in Google. The external packages we create in BzlmodRepoRuleFunction host just one rule each; so instead of storing a rule, we can just store the package and get the rule from there.

This does require adding the rule to the package, which we weren't doing before. This involved a bit of a detour to clean up the RuleFactory#createAndAddRule method, which for some reason has had a much more sensible overload that's package private since time immemorial. This CL simply removes the overload that just gets two fields out of PackageContext.

PiperOrigin-RevId: 387543141