Allow BuildViewTestCase subclasses to inject extra precomputed values
This change allows the subclasses to override the `extraPrecomputedValues` to inject some precomputed values dynamically.

Naively setting the precomputed values in a @Before clause using the Skyframe differencer is insufficient because the "package loading listeners" that trigger an independent PackageLoader run don't get those values. So this had to be routed accordingly, especially in BazelPackageLoader.

PiperOrigin-RevId: 387571896