Change jvm_opts to NestedSet (natively and in Starlark).
When a list of strings is added to command line the command line array is expanded and each string is added to it. However when adding a NestedSet, only a reference to it is added to command line array.

Alternative optimisation would be to also add references to underlying list's array when adding strings to the command line. It is more general optimisation, but the problem is that lists are not that well optimised in keeping the reference to the underlying arrays and it might be impossible to prevent a copy in every location.

The NestedSets are on the other hand already optimised so that copying doesn't happen. But this solution mandates using them in other locations as well - for android_lint jvm_opts and for javac_options.

Change of JavaToolchainProvider's jvm_opt Starlark API to return Depset instead of List could be a breaking change. That happens if you are inspecting the content jvm_opts it requires additional to_list() call. However such cases should be rare.

RELNOTES[INC]: JavaToolchainInfo.jvm_opt returns Depset instead of a list.

PiperOrigin-RevId: 387592165