Split java_plugin into additional library or use the library if available.
This change is needed, because java_plugin will stop working as a dep or runtime_dep of other Java targets.

BEGIN_PUBLIC
Split java_plugin into additional library or use the library if available.

This change is needed, because java_plugin will stop working as a dep or runtime_dep of other Java targets.
END_PUBLIC

PiperOrigin-RevId: 387801584