bzlmod: Use a constant ModuleKey for the root module
(https://github.com/bazelbuild/bazel/issues/13316)

So far we've used <$rootModuleName, Version.EMPTY> as the root ModuleKey. This is a bit unwieldy because we need to pass the root module name everywhere. This CL makes it so that the root module always has the ModuleKey <"", Version.EMPTY>; this does mean that the name of the module key doesn't necessarily match the name of the module, but it makes for cleaner usage everywhere.

PiperOrigin-RevId: 388233328