Rewrite apple_static_library
Two tests were updated to remove their use of the flag --all_incompatible_flags which is no longer needed and will be removed as discussed offline. With that flag the tests fail at this CL because of the use of Starlark implicit outputs. There is an incompatible flag for disabling Starlark implicit outputs, but that flag too will probably be deleted by brandjon@ or ilist@. The use of implicit outputs in apple_static_library is correct.

RELNOTES:none
PiperOrigin-RevId: 388430523