Automated rollback of commit 23f171e2208700dc2f2f65f5ccb32bacd7aec87d.
*** Reason for rollback ***

b/195550666: internal breakage

*** Original change description ***

Avoid touching the filesystem to see what the BUILD file for a Starlark extension was: we already have the package, so we can just read it out.

Genquery remains buggy here.

haxorz@ observed in http://b/123795023#comment6 that we only need to read these paths to determine if the package was defined in a BUILD or BUILD.bazel file. Since the package already knows that, we just ask it.

PiperOrigin-RevId: 389642637