Add the default solib dir to the rpath for shared libs with transitions.
When an executable has dynamically-linked dependencies that have transitions,
shared libraries are looked up under rpaths like this:
    $ORIGIN/(../)*_solib_k8/../../../k8-fastbuild-ST-[0-9a-f]+/bin/_solib_k8

Unless running under the execroot, the transitioned solib directory may not be
available at all; the right files will be present under the default solib
directory, so it should be on the rpath.

Adding the default solib directory to the rpath may cause the wrong version of
a shared library to be loaded, e.g. if it has been compiled in the default
configuration. To prevent this, we also add the transition mnemonic to the
mangled name of the library (adding the default solib last to the rpath won't
really help, because it can legitimately be added first).