BuildEventServiceGrpcClient accepts ClientInterceptor in constructor.
With this refactoring, the `BuildEventServiceGrpcClient` is not concerned with
interpreting command-line options and can operate purely on gRPC domain objects.
The `BazelBuildEventServiceModule` is responsible for preparing the domain objects
from the command-line options.

This refactoring was requested but skipped during PR #13812 because the
BazelBuildEventServiceModuleTest was not yet open-sourced. Now that the test is
open-sourced we can more easily separate (and test) these concerns.

PiperOrigin-RevId: 391347857