Propagate the full workspace status info to cc_library linkstamp.
Minimal patch to fix #1992 and #4863.

Includes both predefined and custom keys as build information
for C++ stamping, treating all unknown keys as strings.

Makes it so that the build-info will be redacted iif both
volatile and non-volatile info is requested in a single header,
since the inputs are now always passed in to figure out the
available keys.