Add a test case for coverage on a Java switch statement.
This acts as a preliminary test for the application of Jacoco's filters.

The java compiler will generate a large number of branches on a string
switch statement (one set for testing against String::hashCode and a
second set as backup) - the majority of these should be filtered as
Jacoco normally would.