Apply Jacoco's filtering to branch coverage analysis.
Jacoco's filters operate on asm.tree.AbstractInsnNode objects. We track
these as we walk through the ASM tree in the same way Jacoco does, apply
the filters to them appropriately, and then map ignored nodes back to
the Instruction objects we already operated on.

We only handle IFilter::ignore for now; this covers the majority of
filters.