Flip incompatible_disallow_resource_jars flag.
The resource_jars is not supported by java_common.compile function. In
order to rewrite java_library to Starlark the call needs to be extended or the
feature disabled. Because the migration is relatively simple I'm opting
to flip the flag and remove it.

RELNOTES[INC]: Flipped --incompatible_disallow_resource_jars (see #13221).