Change DEFAULT_STUB_SHEBANG to use python3 instead of python.
This reflects the change in recent Linux distributions that no longer ship a /usr/bin/python binary.

I'm not sure where this is actually used, now that if I understand correctly the shebang defaults to the Python binary found via auto-detection, but we should probably still fix this path (or remove this default fallback completely?).

Related to #11201. Closes #13851.

PiperOrigin-RevId: 392427248