Flip incompatible_disallow_resource_jars flag.
The resource_jars is not supported by java_common.compile function. In
order to rewrite java_library to Starlark the call needs to be extended or the
feature disabled. Because the migration is relatively simple I'm opting
to flip the flag and remove it.

RELNOTES[INC]: Flipped --incompatible_disallow_resource_jars (see https://github.com/bazelbuild/bazel/issues/13221).

Closes #13891.

PiperOrigin-RevId: 392427418