Allow multiple outputs in SpawnLog
As reported in https://github.com/bazelbuild/bazel/issues/12510 there might be multiple spawns with the same output when flaky tests are executed multiple times.

The fix is to remove the check for presence of duplicate outputs and to extend the sorting algorithm to accept multiple outputs with the same name.

Closes #13650.

PiperOrigin-RevId: 393750094