Add a test cleanup step for unprocessed exception potentially left over in `BugReport`.
When a test fails due to a crash from `BugReport`, an exception will be stored in `BugReport::unprocessedThrowableInTest`. That means that until we clean it up, any following call to `BlazeRuntimeWrapper::executeBuild` will throw that exception, even if that is in the following test.

Add a cleanup step to make sure a test failure cannot taint a following one with an exception stored in `BugReport`.

PiperOrigin-RevId: 393834868