Remove Event's BUILD dependencies on filesystem things
This winds up pulling in all sorts of dependencies (filesystem, profiler) for
what's otherwise a low-level library.

Also, this makes it much more clear that the whole withOutErr thing exists
purely for ferrying action output around the system. Having events be
semi-silently backed by files has been a source of several issues in the past,
hopefully this rephrasing / indirection discourages future uses.

I feel like the code in this area is still somewhat awkward - I tried to look
past that for the most part and stay focused on cutting dependencies.

PiperOrigin-RevId: 393846789