Automated rollback of commit e8ddcf3abc20c1c4dba150267c94d678975d15e4.
*** Reason for rollback ***

Breaks internal tests.

*** Original change description ***

Change Fileset to use providers rather than the FilesetEntry
Starlark value type. This allows aspects to traverse Fileset deps correctly and
allows constraints to propagate as well. This change removes FilesetEntry
completely from the output build graph, and allows all references to this
special type to be removed from the core Bazel code, leaving it as just a
Google-specific extension.

Note that this does not remove the SkyFunction support for Filesets.

PiperOrigin-RevId: 394462193