bzlmod: Consolidate resolution sky functions
(https://github.com/bazelbuild/bazel/issues/13316)

This CL merges DiscoveryFunction and SelectionFunction into one SkyFunction called BazelModuleResolutionFunction. The rationale is that we'll be adding some additional logic at the end of SelectionFunction to prepare for module extension resolution (coming in next CLs), and that logic does not really belong as part of "Selection". The previous split between DiscoveryFunction and SelectionFunction, besides semantics, was to facilitate testing; this is worked around by having a test SkyFunction for discovery within tests.

PiperOrigin-RevId: 395502681