Remove optimization in LibraryToLink that is no longer needed.
This was introduced for Objective-C rules which were adding a lot of repeated
libraries. After Starlarkification of objc_library this is no longer a problem.

The optimization is complex and error prone. It had a bug where it was not
checking that objects were null before falling back to static only.

RELNOTES:none
PiperOrigin-RevId: 395541818