Move starlarkSemantics field from StarlarkCustomCommandLine into the arguments list.
StarlarkCustomCommandLine took 1 pointer more, because it was storing starlarkSemantic unconditionally. The pointer is only needed when MapEach function is used, therefore moving it to the command line.

This reduces retained heap on all Starlark command lines, where map each is not used (and increases it a bit where it is used more than once).

I tried also making a closure for map function with StarlarkSemantics, but that increased the retained heap.

PiperOrigin-RevId: 395646434