Automated rollback of commit d1da69661fb86f3f13d68747ab1a5a5184df1ba6.
*** Reason for rollback ***

This breaks a lot of targets with errors regarding filesets.

*** Original change description ***

Change Fileset to use providers rather than the FilesetEntry
Starlark value type. This allows aspects to traverse Fileset deps correctly and
allows constraints to propagate as well. This change removes FilesetEntry
completely from the output build graph, and allows all references to this
special type to be removed from the core Bazel code, leaving it as just a
Google-specific extension.

Note that this does not remove the SkyFunction support for Filesets.

PiperOrigin-RevId: 395684102