Fix AspectKey for required aspects
This CL removes the inherited propagation information from the AspectKey of the required aspects. This information should not affect the execution of the aspect on a certain target but it only controls its propagation over the dependency graph. Therefore it should not be part of the AspectKey, instead this CL moves the control of the required aspects propagation to the DependencyResolver.

PiperOrigin-RevId: 395730300