Support the Jacoco filter merge operation in Java branch coverage.
Instructions that are indicated to be merged must have their corresponding
coverage expressions (`CovExp`) combined in a pairwise fashion.

e.g. A pair of branch expressions that should be merged:
`BranchExp(a, b)`, `BranchExp(c, d)`

Becomes:

`BranchExp(a | c, b | d)`

This is distinct from the previous merge operation on BranchExp which
simply joined the branches of the two objects (this has been renamed to
concatenate).