Automated rollback of commit f2efbb718aec0a6b47eefc65362205548b33e5a7.
*** Reason for rollback ***

Giving too much access to Starlark rules which leads to additional complexity.

*** Original change description ***

Expose rule kind on Starlark rule context attributes.

The rule kind is needed to implement some checks when porting the rules to Starlark.

The value of a rule kind can already be inspected using an aspect via target.kind, this is just providing a more direct access from the rule implementation.

PiperOrigin-RevId: 396584163