Support replaceBranches operation for Jacoco filters in Java coverage.
Because the filter API takes the replacements as a Set (and the
implementations typically use HashSets), we have to find a way to make
the order of adding the new branches deterministic across multiple
coverage runs so the data can be meaningfully merged. We achieve this by
tracking the order we walk AbstractInsnNodes.

This is not a much used filter but does come in to play with the "when"
feature in Kotlin.

Fixes #12696.

PiperOrigin-RevId: 397054657