Automated rollback of commit 6f5c4bb4d17abd6b9654f255c12775eb35638000.
*** Reason for rollback ***

Fixed handling of `NoSuchFileException`.

*** Original change description ***

Automated rollback of commit 8181f0abe6591bc14957e6a941a6556fb040ca9f.

*** Reason for rollback ***

b/198749645

*** Original change description ***

Add a new `FileSystem` method to get a `SeekableByteChannel` for writing files.

Add `FileSystem::createSeekableChannel` which allows us to create a `SeekableByteChannel` for file writes. The channel is opened in truncated mode, meaning that it is not suitable for reading existing files.

PiperOrigin-RevId: 397097903