added tests, still not sure about disk cache tests
I think shouldUploadLocalResults is poorly named because it is being used on CAS entries, not just action cache results.
But it is the CAS entries for both action inputs and action result outputs.

I'm not sure if downloadBlob needs a similar check or not.
Things seem to work for me without it, but I might not be trying a certain flag setting.