Make bash unit tests infrastructure resilient to `set -eo pipefail`.
When running a bash script with `set -e`, any failed results with the script exiting. In case of the test infrastructure that would mean skipping all post-execution steps (including error reporting) and other test cases. In order to achieve a `try-catch` behavior, we use a pipe which allows us to retain the `set -e` behavior (`set -e` is not effective under operators like `||`) but does not fail the command. We later [use `${PIPESTATUS[0]}` to find the underlying exit code](https://github.com/bazelbuild/bazel/blob/eb349967c8f11bad10db7fe2526bd3c60bca7ee7/src/test/shell/unittest.bash#L716-L721).

This solution works provided `pipefail` is disabled. If it is enabled, the failure code will be carried across the pipeline and `errexit` will make the script fail. Change the logic so that we can temporarily disable `pipefail` only for the execution of subshell which contains the test. Restore the option within the subshell itself and after it so it still applies to the test code and other commands.

PiperOrigin-RevId: 397850055