Add a method to perform the necessary setups for Execution phase.
This new method combines a few setup steps from the various sources into one,
which will come into effect when we successfully merge the Skyframe phases.
This method is to be called before we enter the analysis/execution phase.

Note that the setup is not exhaustive at the current stage, but just enough to handle the "happy path" build without progress reporting and action cache.

RELNOTES: None
PiperOrigin-RevId: 398183724