Don't crash when --compile_one_dependency references a particularly broken package
The rule search logic in CompileOneDependencyTransformer expected that
any targets other than FileTargets, referred to by rules' "direct
compile-time input" LABEL_LIST attributes, would have an "associated
rule".

This isn't true for two kinds of targets: environment groups and
package groups.

Of course it's nonsense for a rule to reference one of these kinds of
targets in an attribute such as "srcs", but malformed user input should
not result in crashes.

RELNOTES: None.
PiperOrigin-RevId: 399194511