Make Bazel's IntelliJ project file more accessible for new developers.
This disables the parts of the build that require an Android SDK / NDK and also two shell tests that currently cause the build to fail on macOS.

This should cause the project file to just work out of the box on most developer machines. Experienced developers can manually modify the project file and enable these tests if they need them.

PiperOrigin-RevId: 399402189