Fix Label() behavior when called with '@repo' parts
The Label() constructor reads its repo mapping from BazelStarlarkContext, which is constructed when the thread is created. This is wrong because the repo mapping it needs is actually the one of the repo where the .bzl file lives. So we do that by having BazelModuleContext store the repo mapping (similar to how it stores the label right now) and return that. Callers who want the repo mapping (not just Label()) now get it from the topmost stack frame rather than from the BazelStarlarkContext, which no longer carries a repo mapping.

This coincidentally also fixes Label() behavior when called from a Bzlmod environment (either extension impl functions, or repo rules; see https://github.com/bazelbuild/bazel/issues/13316), because we simply don't have a BazelStarlarkContext in those cases.

Fixes https://github.com/bazelbuild/bazel/issues/12963.

PiperOrigin-RevId: 401732211