Remove FilesetEntry from Bazel.
It is a Google-internal concept but previously it had to be part of Bazel because it was part of the loading hpase.

Now that https://github.com/bazelbuild/bazel/commit/d1da69661fb86f3f13d68747ab1a5a5184df1ba6 made it unnecessary for the loading phase to know about Filesets, we can remove a bit more knowledge of them from Bazel.

Unfortunately, it's still pretty well embedded in Skyframe but it's still an improvement.

RELNOTES: None.
PiperOrigin-RevId: 402271277