Fake out infrastructure to support (some) cpp-isms in the loading phase
The issue is BazelMockCcSupport (and likely others) don't currently support the
isRealFileSystem=true case. Fake out what is necessary to get by for the time
being while we figure out how to make isRealFileSystem mode work or something
else altogether. We can't unconditionally stage the files because the
isRealFileSystem=false case is used by BuildViewTestCase and seems to mostly
work. We don't just fake isRealFileSystem=false all the time because that needs
a little more investigation and I'd like to get something in to unblock efforts
to open source more tests, which this should do for ones touching just the
loading phase.

Verified that this gets some QueryIntegrationTest methods that were failing due
to missing cpp business to pass.

PiperOrigin-RevId: 402851221