Rewrite `SerializationContextTest` to avoid using mocks.
The current approach of mocking `CodedOutputStream` is dubious because we are verifying calls to `final` methods. In this case, Mockito does not actually verify the specified method call, but instead verifies calls to the methods called within the body of the specified method.

It's not too hard to use real instances for `CodedOutputStream` and `ObjectCodecRegistry` in this test, which is generally better than mocks anyway.

PiperOrigin-RevId: 404089580