Remove test_remote_https_cache.
This test claimed to "Test that if 'https' is provided as a scheme for --remote_cache flag, remote cache works." However, it does not establish that. Failing to read or write to the remote cache simply prints a warning, which does not fail this test.

The HTTPS remote cache cannot work due to at least the following reasons:
1. The remote worker in this test file does not have the HTTP cache enabled.
2. The test remote worker does not support HTTPS over its HTTP cache.
3. Even if the test remote worker did support HTTPS, the HTTP remote cache client does not read --tls_certificate from the command line.